refactor: lookup types and functions
