+++ lib/types/skip-reason.ts
@@ -9,6 +9,7 @@ export enum SkipReason {
   GitPlugin = 'git-plugin',
   Ignored = 'ignored',
   InternalPackage = 'internal-package',
+  InvalidConfig = 'invalid-config',
   InvalidDependencySpecification = 'invalid-dependency-specification',
   InvalidName = 'invalid-name',
   InvalidSha256 = 'invalid-sha256',
@@ -32,11 +33,9 @@ export enum SkipReason {
   UnknownRegistry = 'unknown-registry',
   UnknownVersion = 'unknown-version',
   UnknownVolta = 'unknown-volta',
-  Unknown = 'unknown',
   UnsupportedChartType = 'unsupported-chart-type',
   UnsupportedRemote = 'unsupported-remote',
   UnsupportedUrl = 'unsupported-url',
-  UnsupportedValue = 'unsupported-value',
   UnsupportedVersion = 'unsupported-version',
   Unsupported = 'unsupported',
   UnversionedReference = 'unversioned-reference',
