+++ lib/workers/repository/process/lookup/index.ts
@@ -1,8 +1,4 @@
-import {
-  RenovateConfig,
-  UpdateType,
-  ValidationMessage,
-} from '../../../../config';
+import { ValidationMessage } from '../../../../config';
 import {
   Release,
   getDefaultVersioning,
@@ -14,165 +10,47 @@ import {
 import * as datasourceGitSubmodules from '../../../../datasource/git-submodules';
 import { logger } from '../../../../logger';
 import { getRangeStrategy } from '../../../../manager';
-import { LookupUpdate, RangeConfig } from '../../../../manager/common';
+import { LookupUpdate } from '../../../../manager/common';
 import { SkipReason } from '../../../../types';
 import { clone } from '../../../../util/clone';
 import { applyPackageRules } from '../../../../util/package-rules';
 import * as allVersioning from '../../../../versioning';
-import { FilterConfig, filterVersions } from './filter';
-import { RollbackConfig, getRollbackUpdate } from './rollback';
-
-export interface UpdateResult {
-  sourceDirectory?: string;
-  dockerRepository?: string;
-  dockerRegistry?: string;
-  changelogUrl?: string;
-  dependencyUrl?: string;
-  homepage?: string;
-  deprecationMessage?: string;
-  sourceUrl?: string;
-  skipReason: SkipReason;
-  releases: Release[];
-  fixedVersion?: string;
-  updates: LookupUpdate[];
-  warnings: ValidationMessage[];
-}
-
-export interface LookupUpdateConfig
-  extends RollbackConfig,
-    FilterConfig,
-    RangeConfig,
-    RenovateConfig {
-  separateMinorPatch?: boolean;
-  digestOneAndOnly?: boolean;
-  pinDigests?: boolean;
-  rollbackPrs?: boolean;
-  currentDigest?: string;
-  lockedVersion?: string;
-  vulnerabilityAlert?: boolean;
-  separateMajorMinor?: boolean;
-  separateMultipleMajor?: boolean;
-  datasource: string;
-  depName: string;
-}
-
-function getUpdateType(
-  config: LookupUpdateConfig,
-  currentVersion: string,
-  newVersion: string
-): UpdateType {
-  const { versioning } = config;
-  const version = allVersioning.get(versioning);
-  if (version.getMajor(newVersion) > version.getMajor(currentVersion)) {
-    return 'major';
-  }
-  if (version.getMinor(newVersion) > version.getMinor(currentVersion)) {
-    return 'minor';
-  }
-  if (config.separateMinorPatch) {
-    return 'patch';
-  }
-  if (config.patch.automerge && !config.minor.automerge) {
-    return 'patch';
-  }
-  return 'minor';
-}
-
-function getCurrentVersion(
-  config: LookupUpdateConfig,
-  rangeStrategy: string,
-  latestVersion: string,
-  allVersions: string[]
-): string | null {
-  const { currentValue, lockedVersion, versioning } = config;
-  const version = allVersioning.get(versioning);
-  if (version.isVersion(currentValue)) {
-    return currentValue;
-  }
-  if (version.isSingleVersion(currentValue)) {
-    return currentValue.replace(/=/g, '').trim();
-  }
-  logger.trace(`currentValue ${currentValue} is range`);
-  let useVersions = allVersions.filter((v) => version.matches(v, currentValue));
-  if (latestVersion && version.matches(latestVersion, currentValue)) {
-    useVersions = useVersions.filter(
-      (v) => !version.isGreaterThan(v, latestVersion)
-    );
-  }
-  if (rangeStrategy === 'pin') {
-    return (
-      lockedVersion || version.getSatisfyingVersion(useVersions, currentValue)
-    );
-  }
-  if (rangeStrategy === 'bump') {
-    // Use the lowest version in the current range
-    return version.minSatisfyingVersion(useVersions, currentValue);
-  }
-  // Use the highest version in the current range
-  return version.getSatisfyingVersion(useVersions, currentValue);
-}
-
-function getBucket(
-  config: LookupUpdateConfig,
-  currentVersion: string,
-  newVersion: string,
-  versioning: allVersioning.VersioningApi
-): string {
-  const {
-    separateMajorMinor,
-    separateMultipleMajor,
-    separateMinorPatch,
-  } = config;
-  if (!separateMajorMinor) {
-    return 'latest';
-  }
-  const fromMajor = versioning.getMajor(currentVersion);
-  const toMajor = versioning.getMajor(newVersion);
-  if (fromMajor !== toMajor) {
-    if (separateMultipleMajor) {
-      return `major-${toMajor}`;
-    }
-    return 'major';
-  }
-  if (separateMinorPatch) {
-    if (
-      versioning.getMinor(currentVersion) === versioning.getMinor(newVersion)
-    ) {
-      return 'patch';
-    }
-    return 'minor';
-  }
-  return 'non-major';
-}
+import { getBucket } from './bucket';
+import { LookupUpdateConfig, UpdateResult } from './common';
+import { getCurrentVersion } from './current';
+import { filterVersions } from './filter';
+import { getRollbackUpdate } from './rollback';
+import { getUpdateType } from './update-type';
 
 export async function lookupUpdates(
   inconfig: LookupUpdateConfig
 ): Promise<UpdateResult> {
   let config: LookupUpdateConfig = { ...inconfig };
-  const { depName, currentValue, lockedVersion, vulnerabilityAlert } = config;
+  const {
+    currentDigest,
+    currentValue,
+    datasource,
+    depName,
+    digestOneAndOnly,
+    followTag,
+    lockedVersion,
+    packageFile,
+    pinDigests,
+    rollbackPrs,
+    vulnerabilityAlert,
+  } = config;
   logger.trace({ dependency: depName, currentValue }, 'lookupUpdates');
   // Use the datasource's default versioning if none is configured
   const versioning = allVersioning.get(
-    config.versioning || getDefaultVersioning(config.datasource)
+    config.versioning || getDefaultVersioning(datasource)
   );
   const res: UpdateResult = { updates: [], warnings: [] } as any;
-
-  const isValid = currentValue && versioning.isValid(currentValue);
-  if (!isValid) {
-    res.skipReason = SkipReason.InvalidValue;
-  }
-  // Record if the dep is fixed to a version
-  if (lockedVersion) {
-    res.fixedVersion = lockedVersion;
-  } else if (currentValue && versioning.isSingleVersion(currentValue)) {
-    res.fixedVersion = currentValue.replace(/^=+/, '');
-  }
   // istanbul ignore if
   if (!isGetPkgReleasesConfig(config)) {
-    res.skipReason = SkipReason.Unknown;
+    res.skipReason = SkipReason.InvalidConfig;
     return res;
   }
-
+  const isValid = currentValue && versioning.isValid(currentValue);
   if (isValid) {
     const dependency = clone(await getPkgReleases(config));
     if (!dependency) {
@@ -181,10 +59,7 @@ export async function lookupUpdates(
         depName,
         message: `Failed to look up dependency ${depName}`,
       };
-      logger.debug(
-        { dependency: depName, packageFile: config.packageFile },
-        warning.message
-      );
+      logger.debug({ dependency: depName, packageFile }, warning.message);
       // TODO: return warnings in own field
       res.warnings.push(warning);
       return res;
@@ -215,18 +90,18 @@ export async function lookupUpdates(
     if (allVersions.length === 0) {
       const message = `Found no results from datasource that look like a version`;
       logger.debug({ dependency: depName, result: dependency }, message);
-      if (!config.currentDigest) {
+      if (!currentDigest) {
         return res;
       }
     }
     // Reapply package rules in case we missed something from sourceUrl
     config = applyPackageRules({ ...config, sourceUrl: res.sourceUrl });
-    if (config.followTag) {
-      const taggedVersion = dependency.tags[config.followTag];
+    if (followTag) {
+      const taggedVersion = dependency.tags[followTag];
       if (!taggedVersion) {
         res.warnings.push({
           depName,
-          message: `Can't find version with tag ${config.followTag} for ${depName}`,
+          message: `Can't find version with tag ${followTag} for ${depName}`,
         });
         return res;
       }
@@ -241,7 +116,7 @@ export async function lookupUpdates(
     const allSatisfyingVersions = allVersions.filter((v) =>
       versioning.matches(v.version, currentValue)
     );
-    if (config.rollbackPrs && !allSatisfyingVersions.length) {
+    if (rollbackPrs && !allSatisfyingVersions.length) {
       const rollback = getRollbackUpdate(config, allVersions);
       // istanbul ignore if
       if (!rollback) {
@@ -268,12 +143,14 @@ export async function lookupUpdates(
     const currentVersion =
       getCurrentVersion(
         config,
+        versioning,
         rangeStrategy,
         latestVersion,
         nonDeprecatedVersions
       ) ||
       getCurrentVersion(
         config,
+        versioning,
         rangeStrategy,
         latestVersion,
         allVersions.map((v) => v.version)
@@ -354,7 +231,7 @@ export async function lookupUpdates(
         update.newValue = currentValue;
       }
       if (!update.newValue || update.newValue === currentValue) {
-        if (!config.lockedVersion) {
+        if (!lockedVersion) {
           continue; // eslint-disable-line no-continue
         }
         // istanbul ignore if
@@ -373,7 +250,8 @@ export async function lookupUpdates(
       update.newMajor = versioning.getMajor(newVersion);
       update.newMinor = versioning.getMinor(newVersion);
       update.updateType =
-        update.updateType || getUpdateType(config, currentVersion, newVersion);
+        update.updateType ||
+        getUpdateType(config, versioning, currentVersion, newVersion);
       update.isSingleVersion =
         update.isSingleVersion || !!versioning.isSingleVersion(update.newValue);
       if (!versioning.isVersion(update.newValue)) {
@@ -408,38 +286,41 @@ export async function lookupUpdates(
       res.updates.push(update);
     }
   } else if (!currentValue) {
-    res.skipReason = SkipReason.UnsupportedValue;
+    res.skipReason = SkipReason.InvalidValue;
   } else {
     logger.debug(`Dependency ${depName} has unsupported value ${currentValue}`);
-    if (!config.pinDigests && !config.currentDigest) {
-      res.skipReason = SkipReason.UnsupportedValue;
+    if (!pinDigests && !currentDigest) {
+      res.skipReason = SkipReason.InvalidValue;
     } else {
       delete res.skipReason;
     }
   }
+  // Record if the dep is fixed to a version
+  if (lockedVersion) {
+    res.fixedVersion = lockedVersion;
+  } else if (currentValue && versioning.isSingleVersion(currentValue)) {
+    res.fixedVersion = currentValue.replace(/^=+/, '');
+  }
   // Add digests if necessary
-  if (config.newDigest || supportsDigests(config)) {
-    if (
-      config.currentDigest &&
-      config.datasource !== datasourceGitSubmodules.id
-    ) {
-      if (!config.digestOneAndOnly || !res.updates.length) {
+  if (supportsDigests(config)) {
+    if (currentDigest && datasource !== datasourceGitSubmodules.id) {
+      if (!digestOneAndOnly || !res.updates.length) {
         // digest update
         res.updates.push({
           updateType: 'digest',
-          newValue: config.currentValue,
+          newValue: currentValue,
         });
       }
-    } else if (config.pinDigests) {
+    } else if (pinDigests) {
       // Create a pin only if one doesn't already exists
       if (!res.updates.some((update) => update.updateType === 'pin')) {
         // pin digest
         res.updates.push({
           updateType: 'pin',
-          newValue: config.currentValue,
+          newValue: currentValue,
         });
       }
-    } else if (config.datasource === datasourceGitSubmodules.id) {
+    } else if (datasource === datasourceGitSubmodules.id) {
       const dependency = clone(await getPkgReleases(config));
       if (dependency?.releases[0]?.version) {
         res.updates.push({
@@ -459,7 +340,7 @@ export async function lookupUpdates(
     }
     // update digest for all
     for (const update of res.updates) {
-      if (config.pinDigests || config.currentDigest) {
+      if (pinDigests || currentDigest) {
         update.newDigest =
           update.newDigest || (await getDigest(config, update.newValue));
         if (update.newDigest) {
@@ -480,9 +361,9 @@ export async function lookupUpdates(
     .filter((update) => update.newDigest !== null)
     .filter(
       (update) =>
-        update.newValue !== config.currentValue ||
+        update.newValue !== currentValue ||
         update.isLockfileUpdate ||
-        (update.newDigest && !update.newDigest.startsWith(config.currentDigest))
+        (update.newDigest && !update.newDigest.startsWith(currentDigest))
     );
   if (res.updates.some((update) => update.updateType === 'pin')) {
     for (const update of res.updates) {
